add iterative trimming with parallel analysis
